moved internal macros into separate crate
