+++ mlx-internal-macros/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "mlx-internal-macros"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2.0.60", features = ["full"] }
+quote = "1.0"
+darling = "0.20"
+proc-macro2 = "1.0"
