+++ mlx-internal-macros/src/generate_builder.rs
