+++ mlx-nn/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 mlx-rs.workspace = true
 mlx-macros.workspace = true
+mlx-internal-macros.workspace = true
 thiserror.workspace = true
 
 [dev-dependencies]
