+++ mlx-rs/src/fft/fftn.rs
@@ -1,4 +1,4 @@
-use mlx_macros::default_device;
+use mlx_internal_macros::default_device;
 
 use crate::{array::Array, error::Exception, stream::StreamOrDevice, utils::IntoOption, Stream};
