+++ mlx-rs/src/ops/indexing/mod.rs
@@ -101,7 +101,7 @@ use std::{
     ops::{Bound, RangeBounds},
 };
 
-use mlx_macros::default_device;
+use mlx_internal_macros::default_device;
 
 use crate::{error::Exception, Array, Stream, StreamOrDevice};
