+++ mlx-rs/src/ops/sort.rs
@@ -1,6 +1,6 @@
 //! Implements bindings for the sorting ops.
 
-use mlx_macros::default_device;
+use mlx_internal_macros::default_device;
 
 use crate::{error::Exception, Array, Stream, StreamOrDevice};
